requirements.txt
# Core FastAPI and web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client and async support
httpx>=0.25.2
aiohttp>=3.9.1
requests>=2.31.0

# Groq LLM client
groq>=0.4.2
langchain-groq>=0.1.0

# Vector database and embeddings
chromadb>=0.4.18
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
transformers>=4.36.2
torch>=2.7.1

# Real-time knowledge search
# langchain-tavily is specified in LangChain section below

# Translation and language processing
sacremoses>=0.1.1
tokenizers>=0.15.0

# Medical and scientific libraries
scikit-learn>=1.3.2
numpy>=1.24.3
pandas>=2.1.4
scipy>=1.11.4

# Database and caching
redis>=5.0.1
sqlalchemy>=2.0.23
alembic>=1.13.1
psycopg2-binary>=2.9.9  # PostgreSQL adapter
asyncpg>=0.29.0  # Async PostgreSQL support

# Security and authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Monitoring and logging
prometheus-client>=0.19.0
structlog>=23.2.0
loguru>=0.7.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development and linting
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1

# Utilities
python-dotenv>=1.0.0
click>=8.1.7
tqdm>=4.66.1
rich>=13.7.0

# Medical knowledge processing
beautifulsoup4>=4.12.2
lxml>=4.9.3
markdown>=3.5.1
python-docx>=1.1.0
PyPDF2>=3.0.1

# Performance and optimization
orjson>=3.9.10
ujson>=5.8.0
msgpack>=1.0.7

# Docker and deployment
docker>=6.1.3
kubernetes>=28.1.0

# LangChain and Tavily integration
langchain>=0.3.20
langchain-core>=0.3.15
langchain-groq>=0.1.0
langchain-tavily>=0.2.11

# MCP (Model Context Protocol) - optional
# mcp>=1.0.0

# XAI Grok API client - optional
# xai>=0.1.0

# Optional: GPU support (uncomment if using CUDA)
# torch==2.1.2+cu118 --index-url https://download.pytorch.org/whl/cu118
# faiss-gpu==1.7.4 